remove PreventBakeTagComponent
